@@ -55,7 +55,12 @@ import XCTest
         _ = try await ref.delete()
       } catch {
         caughtError = true
-        XCTAssertEqual((error as NSError).code, StorageErrorCode.objectNotFound.rawValue)
+        let nsError = error as NSError
+        XCTAssertEqual(nsError.code, StorageErrorCode.objectNotFound.rawValue)
+        XCTAssertEqual(nsError.userInfo["ResponseErrorCode"] as? Int, 404)
+        let underlyingError = try XCTUnwrap(nsError.userInfo[NSUnderlyingErrorKey] as? NSError)
+        XCTAssertEqual(underlyingError.code, 404)
+        XCTAssertEqual(underlyingError.domain, "com.google.HTTPStatus")
       }
       XCTAssertTrue(caughtError)
     }
@@ -132,8 +137,11 @@ import XCTest
       do {
         _ = try await ref.putFileAsync(from: fileURL)
         XCTFail("Unexpected success from putFile of a directory")
-      } catch StorageError.unknown {
-        XCTAssertTrue(true)
+      } catch let StorageError.unknown(reason) {
+        XCTAssertTrue(reason.starts(with: "File at URL:"))
+        XCTAssertTrue(reason.hasSuffix(
+          "is not reachable. Ensure file URL is not a directory, symbolic link, or invalid url."
+        ))
       } catch {
         XCTFail("error failed to convert to StorageError.unknown")
       }
