@@ -18,9 +18,6 @@
 
 #import <XCTest/XCTest.h>
 
-#import "Firestore/Source/API/FIRFilter+Internal.h"
-#import "Firestore/Source/API/FIRQuery+Internal.h"
-
 #import "Firestore/Example/Tests/Util/FSTEventAccumulator.h"
 #import "Firestore/Example/Tests/Util/FSTHelpers.h"
 #import "Firestore/Example/Tests/Util/FSTIntegrationTestCase.h"
@@ -884,16 +881,8 @@
   [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter1]
                      matchesResult:@[ @"doc1", @"doc2", @"doc4", @"doc5" ]];
 
-  // with one inequality: a>2 || b==1.
-  FIRFilter *filter2 = [FIRFilter orFilterWithFilters:@[
-    [FIRFilter filterWhereField:@"a" isGreaterThan:@2], [FIRFilter filterWhereField:@"b"
-                                                                          isEqualTo:@1]
-  ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter2]
-                     matchesResult:@[ @"doc5", @"doc2", @"doc3" ]];
-
   // (a==1 && b==0) || (a==3 && b==2)
-  FIRFilter *filter3 = [FIRFilter orFilterWithFilters:@[
+  FIRFilter *filter2 = [FIRFilter orFilterWithFilters:@[
     [FIRFilter andFilterWithFilters:@[
       [FIRFilter filterWhereField:@"a" isEqualTo:@1], [FIRFilter filterWhereField:@"b" isEqualTo:@0]
     ]],
@@ -901,20 +890,20 @@
       [FIRFilter filterWhereField:@"a" isEqualTo:@3], [FIRFilter filterWhereField:@"b" isEqualTo:@2]
     ]]
   ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter3]
+  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter2]
                      matchesResult:@[ @"doc1", @"doc3" ]];
 
   // a==1 && (b==0 || b==3).
-  FIRFilter *filter4 = [FIRFilter andFilterWithFilters:@[
+  FIRFilter *filter3 = [FIRFilter andFilterWithFilters:@[
     [FIRFilter filterWhereField:@"a" isEqualTo:@1], [FIRFilter orFilterWithFilters:@[
       [FIRFilter filterWhereField:@"b" isEqualTo:@0], [FIRFilter filterWhereField:@"b" isEqualTo:@3]
     ]]
   ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter4]
+  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter3]
                      matchesResult:@[ @"doc1", @"doc4" ]];
 
   // (a==2 || b==2) && (a==3 || b==3)
-  FIRFilter *filter5 = [FIRFilter andFilterWithFilters:@[
+  FIRFilter *filter4 = [FIRFilter andFilterWithFilters:@[
     [FIRFilter orFilterWithFilters:@[
       [FIRFilter filterWhereField:@"a" isEqualTo:@2], [FIRFilter filterWhereField:@"b" isEqualTo:@2]
     ]],
@@ -922,51 +911,74 @@
       [FIRFilter filterWhereField:@"a" isEqualTo:@3], [FIRFilter filterWhereField:@"b" isEqualTo:@3]
     ]]
   ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter5] matchesResult:@[ @"doc3" ]];
+  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter4] matchesResult:@[ @"doc3" ]];
+
+  // Test with limits without orderBy (the __name__ ordering is the tie breaker).
+  FIRFilter *filter5 = [FIRFilter orFilterWithFilters:@[
+    [FIRFilter filterWhereField:@"a" isEqualTo:@2], [FIRFilter filterWhereField:@"b" isEqualTo:@1]
+  ]];
+  [self checkOnlineAndOfflineQuery:[[collRef queryWhereFilter:filter5] queryLimitedTo:1]
+                     matchesResult:@[ @"doc2" ]];
+}
+
+- (void)testOrQueriesWithCompositeIndexes {
+  // TODO(orquery): Enable this test against production when possible.
+  XCTSkipIf(![FSTIntegrationTestCase isRunningAgainstEmulator],
+            "Skip this test if running against production because order-by-equality is not "
+            "supported yet.");
+
+  FIRCollectionReference *collRef = [self collectionRefWithDocuments:@{
+    @"doc1" : @{@"a" : @1, @"b" : @0},
+    @"doc2" : @{@"a" : @2, @"b" : @1},
+    @"doc3" : @{@"a" : @3, @"b" : @2},
+    @"doc4" : @{@"a" : @1, @"b" : @3},
+    @"doc5" : @{@"a" : @1, @"b" : @1}
+  }];
+
+  // with one inequality: a>2 || b==1.
+  FIRFilter *filter1 = [FIRFilter orFilterWithFilters:@[
+    [FIRFilter filterWhereField:@"a" isGreaterThan:@2], [FIRFilter filterWhereField:@"b"
+                                                                          isEqualTo:@1]
+  ]];
+  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter1]
+                     matchesResult:@[ @"doc5", @"doc2", @"doc3" ]];
 
   // Test with limits (implicit order by ASC): (a==1) || (b > 0) LIMIT 2
-  FIRFilter *filter6 = [FIRFilter orFilterWithFilters:@[
+  FIRFilter *filter2 = [FIRFilter orFilterWithFilters:@[
     [FIRFilter filterWhereField:@"a" isEqualTo:@1], [FIRFilter filterWhereField:@"b"
                                                                   isGreaterThan:@0]
   ]];
-  [self checkOnlineAndOfflineQuery:[[collRef queryWhereFilter:filter6] queryLimitedTo:2]
+  [self checkOnlineAndOfflineQuery:[[collRef queryWhereFilter:filter2] queryLimitedTo:2]
                      matchesResult:@[ @"doc1", @"doc2" ]];
 
   // Test with limits (explicit order by): (a==1) || (b > 0) LIMIT_TO_LAST 2
   // Note: The public query API does not allow implicit ordering when limitToLast is used.
-  FIRFilter *filter7 = [FIRFilter orFilterWithFilters:@[
+  FIRFilter *filter3 = [FIRFilter orFilterWithFilters:@[
     [FIRFilter filterWhereField:@"a" isEqualTo:@1], [FIRFilter filterWhereField:@"b"
                                                                   isGreaterThan:@0]
   ]];
-  [self checkOnlineAndOfflineQuery:[[[collRef queryWhereFilter:filter7] queryLimitedToLast:2]
+  [self checkOnlineAndOfflineQuery:[[[collRef queryWhereFilter:filter3] queryLimitedToLast:2]
                                        queryOrderedByField:@"b"]
                      matchesResult:@[ @"doc3", @"doc4" ]];
 
   // Test with limits (explicit order by ASC): (a==2) || (b == 1) ORDER BY a LIMIT 1
-  FIRFilter *filter8 = [FIRFilter orFilterWithFilters:@[
+  FIRFilter *filter4 = [FIRFilter orFilterWithFilters:@[
     [FIRFilter filterWhereField:@"a" isEqualTo:@2], [FIRFilter filterWhereField:@"b" isEqualTo:@1]
   ]];
-  [self checkOnlineAndOfflineQuery:[[[collRef queryWhereFilter:filter8] queryLimitedTo:1]
+  [self checkOnlineAndOfflineQuery:[[[collRef queryWhereFilter:filter4] queryLimitedTo:1]
                                        queryOrderedByField:@"a"]
                      matchesResult:@[ @"doc5" ]];
 
   // Test with limits (explicit order by DESC): (a==2) || (b == 1) ORDER BY a LIMIT_TO_LAST 1
-  FIRFilter *filter9 = [FIRFilter orFilterWithFilters:@[
+  FIRFilter *filter5 = [FIRFilter orFilterWithFilters:@[
     [FIRFilter filterWhereField:@"a" isEqualTo:@2], [FIRFilter filterWhereField:@"b" isEqualTo:@1]
   ]];
-  [self checkOnlineAndOfflineQuery:[[[collRef queryWhereFilter:filter9] queryLimitedToLast:1]
+  [self checkOnlineAndOfflineQuery:[[[collRef queryWhereFilter:filter5] queryLimitedToLast:1]
                                        queryOrderedByField:@"a"]
-                     matchesResult:@[ @"doc2" ]];
-
-  // Test with limits without orderBy (the __name__ ordering is the tie breaker).
-  FIRFilter *filter10 = [FIRFilter orFilterWithFilters:@[
-    [FIRFilter filterWhereField:@"a" isEqualTo:@2], [FIRFilter filterWhereField:@"b" isEqualTo:@1]
-  ]];
-  [self checkOnlineAndOfflineQuery:[[collRef queryWhereFilter:filter10] queryLimitedTo:1]
                      matchesResult:@[ @"doc2" ]];
 }
 
-- (void)testOrQueriesWithInAndNotIn {
+- (void)testOrQueriesWithIn {
   FIRCollectionReference *collRef = [self collectionRefWithDocuments:@{
     @"doc1" : @{@"a" : @1, @"b" : @0},
     @"doc2" : @{@"b" : @1},
@@ -977,19 +989,35 @@
   }];
 
   // a==2 || b in [2,3]
-  FIRFilter *filter1 = [FIRFilter orFilterWithFilters:@[
+  FIRFilter *filter = [FIRFilter orFilterWithFilters:@[
     [FIRFilter filterWhereField:@"a" isEqualTo:@2], [FIRFilter filterWhereField:@"b" in:@[ @2, @3 ]]
   ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter1]
+  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter]
                      matchesResult:@[ @"doc3", @"doc4", @"doc6" ]];
+}
+
+- (void)testOrQueriesWithNotIn {
+  // TODO(orquery): Enable this test against production when possible.
+  XCTSkipIf(![FSTIntegrationTestCase isRunningAgainstEmulator],
+            "Skip this test if running against production because it results in a 'missing index' "
+            "error. The Firestore Emulator, however, does serve these queries");
+
+  FIRCollectionReference *collRef = [self collectionRefWithDocuments:@{
+    @"doc1" : @{@"a" : @1, @"b" : @0},
+    @"doc2" : @{@"b" : @1},
+    @"doc3" : @{@"a" : @3, @"b" : @2},
+    @"doc4" : @{@"a" : @1, @"b" : @3},
+    @"doc5" : @{@"a" : @1},
+    @"doc6" : @{@"a" : @2}
+  }];
 
   // a==2 || b not-in [2,3]
   // Has implicit orderBy b.
-  FIRFilter *filter2 = [FIRFilter orFilterWithFilters:@[
+  FIRFilter *filter = [FIRFilter orFilterWithFilters:@[
     [FIRFilter filterWhereField:@"a" isEqualTo:@2], [FIRFilter filterWhereField:@"b"
                                                                           notIn:@[ @2, @3 ]]
   ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter2]
+  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter]
                      matchesResult:@[ @"doc1", @"doc2" ]];
 }
 
@@ -1021,6 +1049,10 @@
 }
 
 - (void)testMultipleInOps {
+  // TODO(orquery): Enable this test against production when possible.
+  XCTSkipIf(![FSTIntegrationTestCase isRunningAgainstEmulator],
+            "Skip this test if running against production because it's not yet supported.");
+
   FIRCollectionReference *collRef = [self collectionRefWithDocuments:@{
     @"doc1" : @{@"a" : @1, @"b" : @0},
     @"doc2" : @{@"b" : @1},
@@ -1035,66 +1067,24 @@
     [FIRFilter filterWhereField:@"a" in:@[ @2, @3 ]], [FIRFilter filterWhereField:@"b"
                                                                                in:@[ @0, @2 ]]
   ]];
-  [self checkOnlineAndOfflineQuery:[[collRef queryWhereFilter:filter1] queryOrderedByField:@"a"]
-                     matchesResult:@[ @"doc1", @"doc6", @"doc3" ]];
+  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter1]
+                     matchesResult:@[ @"doc1", @"doc3", @"doc6" ]];
 
-  // Two IN operations on different fields with conjunction.
-  FIRFilter *filter2 = [FIRFilter andFilterWithFilters:@[
-    [FIRFilter filterWhereField:@"a" in:@[ @2, @3 ]], [FIRFilter filterWhereField:@"b"
-                                                                               in:@[ @0, @2 ]]
-  ]];
-  [self checkOnlineAndOfflineQuery:[[collRef queryWhereFilter:filter2] queryOrderedByField:@"a"]
-                     matchesResult:@[ @"doc3" ]];
-
-  // Two IN operations on the same field.
-  // a IN [1,2,3] && a IN [0,1,4] should result in "a==1".
-  FIRFilter *filter3 = [FIRFilter andFilterWithFilters:@[
-    [FIRFilter filterWhereField:@"a" in:@[ @1, @2, @3 ]],
-    [FIRFilter filterWhereField:@"a" in:@[ @0, @1, @4 ]]
-  ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter3]
-                     matchesResult:@[ @"doc1", @"doc4", @"doc5" ]];
-
-  // a IN [2,3] && a IN [0,1,4] is never true and so the result should be an empty set.
-  FIRFilter *filter4 = [FIRFilter andFilterWithFilters:@[
-    [FIRFilter filterWhereField:@"a" in:@[ @2, @3 ]], [FIRFilter filterWhereField:@"a"
-                                                                               in:@[ @0, @1, @4 ]]
-  ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter4] matchesResult:@[]];
-
+  // Two IN operations on the same field with disjunction.
   // a IN [0,3] || a IN [0,2] should union them (similar to: a IN [0,2,3]).
-  FIRFilter *filter5 = [FIRFilter orFilterWithFilters:@[
+  FIRFilter *filter2 = [FIRFilter orFilterWithFilters:@[
     [FIRFilter filterWhereField:@"a" in:@[ @0, @3 ]], [FIRFilter filterWhereField:@"a"
                                                                                in:@[ @0, @2 ]]
   ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter5]
+  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter2]
                      matchesResult:@[ @"doc3", @"doc6" ]];
-
-  // Nested composite filter on the same field.
-  FIRFilter *filter6 = [FIRFilter andFilterWithFilters:@[
-    [FIRFilter filterWhereField:@"a" in:@[ @1, @3 ]], [FIRFilter orFilterWithFilters:@[
-      [FIRFilter filterWhereField:@"a" in:@[ @0, @2 ]], [FIRFilter andFilterWithFilters:@[
-        [FIRFilter filterWhereField:@"b" isGreaterThanOrEqualTo:@1],
-        [FIRFilter filterWhereField:@"a" in:@[ @1, @3 ]]
-      ]]
-    ]]
-  ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter6]
-                     matchesResult:@[ @"doc3", @"doc4" ]];
-
-  // Nested composite filter on different fields.
-  FIRFilter *filter7 = [FIRFilter andFilterWithFilters:@[
-    [FIRFilter filterWhereField:@"b" in:@[ @0, @3 ]], [FIRFilter orFilterWithFilters:@[
-      [FIRFilter filterWhereField:@"b" in:@[ @1 ]], [FIRFilter andFilterWithFilters:@[
-        [FIRFilter filterWhereField:@"b" in:@[ @2, @3 ]], [FIRFilter filterWhereField:@"a"
-                                                                                   in:@[ @1, @3 ]]
-      ]]
-    ]]
-  ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter7] matchesResult:@[ @"doc4" ]];
 }
 
-- (void)testUseInWithArrayContainsAny {
+- (void)testUsingInWithArrayContainsAny {
+  // TODO(orquery): Enable this test against production when possible.
+  XCTSkipIf(![FSTIntegrationTestCase isRunningAgainstEmulator],
+            "Skip this test if running against production because it's not yet supported.");
+
   FIRCollectionReference *collRef = [self collectionRefWithDocuments:@{
     @"doc1" : @{@"a" : @1, @"b" : @[ @0 ]},
     @"doc2" : @{@"b" : @[ @1 ]},
@@ -1111,30 +1101,15 @@
   [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter1]
                      matchesResult:@[ @"doc1", @"doc3", @"doc4", @"doc6" ]];
 
-  FIRFilter *filter2 = [FIRFilter andFilterWithFilters:@[
-    [FIRFilter filterWhereField:@"a" in:@[ @2, @3 ]], [FIRFilter filterWhereField:@"b"
-                                                                 arrayContainsAny:@[ @0, @7 ]]
-  ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter2] matchesResult:@[ @"doc3" ]];
-
-  FIRFilter *filter3 = [FIRFilter orFilterWithFilters:@[
+  FIRFilter *filter2 = [FIRFilter orFilterWithFilters:@[
     [FIRFilter andFilterWithFilters:@[
       [FIRFilter filterWhereField:@"a" in:@[ @2, @3 ]], [FIRFilter filterWhereField:@"c"
                                                                           isEqualTo:@10]
     ]],
     [FIRFilter filterWhereField:@"b" arrayContainsAny:@[ @0, @7 ]]
   ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter3]
+  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter2]
                      matchesResult:@[ @"doc1", @"doc3", @"doc4" ]];
-
-  FIRFilter *filter4 = [FIRFilter andFilterWithFilters:@[
-    [FIRFilter filterWhereField:@"a" in:@[ @2, @3 ]], [FIRFilter orFilterWithFilters:@[
-      [FIRFilter filterWhereField:@"b" arrayContainsAny:@[ @0, @7 ]],
-      [FIRFilter filterWhereField:@"c" isEqualTo:@20]
-    ]]
-  ]];
-  [self checkOnlineAndOfflineQuery:[collRef queryWhereFilter:filter4]
-                     matchesResult:@[ @"doc3", @"doc6" ]];
 }
 
 - (void)testUseInWithArrayContains {
@@ -1179,6 +1154,11 @@
 }
 
 - (void)testOrderByEquality {
+  // TODO(orquery): Enable this test against production when possible.
+  XCTSkipIf(![FSTIntegrationTestCase isRunningAgainstEmulator],
+            "Skip this test if running against production because order-by-equality is not "
+            "supported yet.");
+
   FIRCollectionReference *collRef = [self collectionRefWithDocuments:@{
     @"doc1" : @{@"a" : @1, @"b" : @[ @0 ]},
     @"doc2" : @{@"b" : @[ @1 ]},
