@@ -18,6 +18,8 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+@protocol FIRLocalCacheSettings;
+
 /** Used to set on-disk cache size to unlimited. Garbage collection will not run. */
 FOUNDATION_EXTERN const int64_t
     kFIRFirestoreCacheSizeUnlimited NS_SWIFT_NAME(FirestoreCacheSizeUnlimited);
@@ -34,7 +36,7 @@ NS_SWIFT_NAME(FirestoreSettings)
 - (instancetype)init NS_DESIGNATED_INITIALIZER;
 
 /** The hostname to connect to. */
-@property(nonatomic, copy) NSString *host;
+@property(nonatomic, copy) NSString* host;
 
 /** Whether to use SSL when connecting. */
 @property(nonatomic, getter=isSSLEnabled) BOOL sslEnabled;
@@ -45,19 +47,41 @@ NS_SWIFT_NAME(FirestoreSettings)
  */
 @property(nonatomic, strong) dispatch_queue_t dispatchQueue;
 
-/** Set to false to disable local persistent storage. */
-@property(nonatomic, getter=isPersistenceEnabled) BOOL persistenceEnabled;
+/**
+ * NOTE: This field will be deprecated in a future major release. Use the `cacheSettings` field
+ * instead to specify cache type, and other cache configurations.
+ *
+ * Set to false to disable local persistent storage.
+ */
+@property(nonatomic, getter=isPersistenceEnabled) BOOL persistenceEnabled DEPRECATED_MSG_ATTRIBUTE(
+    "This field is deprecated. Use `cacheSettings` instead.");
 
 @property(nonatomic, getter=isMemoryLruGCEnabled) BOOL memoryLruGCEnabled;
 
 /**
+ * NOTE: This field will be deprecated in a future major release. Use the `cacheSettings` field
+ * instead to specify cache size, and other cache configurations.
+ *
  * Sets the cache size threshold above which the SDK will attempt to collect least-recently-used
  * documents. The size is not a guarantee that the cache will stay below that size, only that if
  * the cache exceeds the given size, cleanup will be attempted. Cannot be set lower than 1MB.
  *
  * Set to `FirestoreCacheSizeUnlimited` to disable garbage collection entirely.
  */
-@property(nonatomic, assign) int64_t cacheSizeBytes;
+@property(nonatomic, assign) int64_t cacheSizeBytes DEPRECATED_MSG_ATTRIBUTE(
+    "This field is deprecated. Use `cacheSettings` instead.");
+
+/**
+ * Specifies the cache used by the SDK. Available options are `PersistentCacheSettings`
+ * and `MemoryCacheSettings`, each with different configuration options.
+ *
+ * When unspecified, `PersistentCacheSettings` will be used by default.
+ *
+ * NOTE: setting this field and `cacheSizeBytes` or `persistenceEnabled` at the same time will throw
+ * an exception during SDK initialization. Instead, use the configuration in
+ * the `PersistentCacheSettings` object to specify the cache size.
+ */
+@property(nonatomic, strong) id<FIRLocalCacheSettings, NSObject> cacheSettings;
 
 @end
 
