@@ -15,8 +15,12 @@
  */
 
 #include "Firestore/core/src/api/settings.h"
+#include <cstddef>
+#include <memory>
 
+#include "Firestore/core/src/util/hard_assert.h"
 #include "Firestore/core/src/util/hashing.h"
+#include "absl/memory/memory.h"
 
 namespace firebase {
 namespace firestore {
@@ -29,16 +33,187 @@ constexpr bool Settings::DefaultLruGcEnabled;
 constexpr int64_t Settings::DefaultCacheSizeBytes;
 constexpr int64_t Settings::MinimumCacheSizeBytes;
 
+Settings::Settings(const Settings& other)
+    : host_(other.host_),
+      ssl_enabled_(other.ssl_enabled_),
+      persistence_enabled_(other.persistence_enabled_),
+      cache_size_bytes_(other.cache_size_bytes_) {
+  if (other.cache_settings_ != nullptr) {
+    cache_settings_ = CopyCacheSettings(*other.cache_settings_);
+  }
+}
+
+Settings& Settings::operator=(const Settings& other) {
+  if (this == &other) {
+    return *this;
+  }
+
+  host_ = other.host_;
+  ssl_enabled_ = other.ssl_enabled_;
+  persistence_enabled_ = other.persistence_enabled_;
+  cache_size_bytes_ = other.cache_size_bytes_;
+  if (other.cache_settings_ != nullptr) {
+    cache_settings_ = CopyCacheSettings(*other.cache_settings_);
+  }
+  return *this;
+}
+
+std::unique_ptr<LocalCacheSettings> Settings::CopyCacheSettings(
+    const LocalCacheSettings& settings) {
+  if (settings.kind_ == LocalCacheSettings::Kind::kMemory) {
+    return absl::make_unique<MemoryCacheSettings>(
+        static_cast<const MemoryCacheSettings&>(settings));
+  } else if (settings.kind_ == LocalCacheSettings::Kind::kPersistent) {
+    return absl::make_unique<PersistentCacheSettings>(
+        static_cast<const PersistentCacheSettings&>(settings));
+  }
+  UNREACHABLE();
+}
+
 size_t Settings::Hash() const {
+  if (cache_settings_ == nullptr) {
+    return util::Hash(host_, ssl_enabled_, persistence_enabled_,
+                      cache_size_bytes_);
+  }
+
   return util::Hash(host_, ssl_enabled_, persistence_enabled_,
-                    cache_size_bytes_);
+                    cache_size_bytes_, *cache_settings_);
 }
 
 bool operator==(const Settings& lhs, const Settings& rhs) {
-  return lhs.host_ == rhs.host_ && lhs.ssl_enabled_ == rhs.ssl_enabled_ &&
-         lhs.persistence_enabled_ == rhs.persistence_enabled_ &&
-         lhs.lru_gc_enabled_ == rhs.lru_gc_enabled_ &&
-         lhs.cache_size_bytes_ == rhs.cache_size_bytes_;
+  bool eq = lhs.host_ == rhs.host_ && lhs.ssl_enabled_ == rhs.ssl_enabled_ &&
+            lhs.persistence_enabled_ == rhs.persistence_enabled_ &&
+            lhs.cache_size_bytes_ == rhs.cache_size_bytes_;
+  if (!eq) {
+    return eq;
+  }
+
+  if (lhs.cache_settings_ == nullptr && rhs.cache_settings_ == nullptr) {
+    return eq;
+  }
+
+  if (lhs.cache_settings_ == nullptr || rhs.cache_settings_ == nullptr) {
+    return false;
+  }
+  return *(lhs.cache_settings_) == *(rhs.cache_settings_);
+}
+
+bool operator!=(const Settings& lhs, const Settings& rhs) {
+  return !(lhs == rhs);
+}
+
+bool operator==(const LocalCacheSettings& lhs, const LocalCacheSettings& rhs) {
+  if (lhs.kind() != rhs.kind()) {
+    return false;
+  }
+
+  if (lhs.kind() == LocalCacheSettings::Kind::kMemory) {
+    return static_cast<const MemoryCacheSettings&>(lhs) ==
+           static_cast<const MemoryCacheSettings&>(rhs);
+  }
+
+  if (lhs.kind() == LocalCacheSettings::Kind::kPersistent) {
+    return static_cast<const PersistentCacheSettings&>(lhs) ==
+           static_cast<const PersistentCacheSettings&>(rhs);
+  }
+  UNREACHABLE();
+}
+
+bool operator!=(const LocalCacheSettings& lhs, const LocalCacheSettings& rhs) {
+  return !(lhs == rhs);
+}
+
+size_t MemoryCacheSettings::Hash() const {
+  return util::Hash(kind_);
+}
+
+size_t PersistentCacheSettings::Hash() const {
+  return util::Hash(kind_, size_bytes_);
+}
+
+bool operator==(const MemoryCacheSettings& lhs,
+                const MemoryCacheSettings& rhs) {
+  return lhs.kind() == rhs.kind();
+}
+bool operator!=(const MemoryCacheSettings& lhs,
+                const MemoryCacheSettings& rhs) {
+  return !(lhs == rhs);
+}
+
+bool operator==(const PersistentCacheSettings& lhs,
+                const PersistentCacheSettings& rhs) {
+  return lhs.kind() == rhs.kind() && lhs.size_bytes() == rhs.size_bytes();
+}
+
+bool operator!=(const PersistentCacheSettings& lhs,
+                const PersistentCacheSettings& rhs) {
+  return !(lhs == rhs);
+}
+
+void Settings::set_persistence_enabled(bool value) {
+  HARD_ASSERT(cache_settings_ == nullptr,
+              "Cannot change persistence when "
+              "local cache settings is already specified. Instead, specify "
+              "persistence as part of local cache settings.");
+  persistence_enabled_ = value;
+}
+
+bool Settings::persistence_enabled() const {
+  return (cache_settings_ &&
+          cache_settings_->kind() == LocalCacheSettings::Kind::kPersistent) ||
+         persistence_enabled_;
+}
+
+void Settings::set_cache_size_bytes(int64_t value) {
+  HARD_ASSERT(cache_settings_ == nullptr,
+              "Cannot change cache size when "
+              "local cache settings is already specified. Instead, specify "
+              "cache size as part of local cache settings.");
+  cache_size_bytes_ = value;
+}
+
+int64_t Settings::cache_size_bytes() const {
+  if (cache_settings_) {
+    if (cache_settings_->kind() == api::LocalCacheSettings::Kind::kPersistent) {
+      return static_cast<const PersistentCacheSettings*>(cache_settings_.get())
+          ->size_bytes_;
+    } else {
+      return CacheSizeUnlimited;
+    }
+  }
+  return cache_size_bytes_;
+}
+
+bool Settings::gc_enabled() const {
+  if (cache_settings_) {
+    return cache_settings_->kind_ == LocalCacheSettings::Kind::kPersistent &&
+           static_cast<PersistentCacheSettings*>(cache_settings_.get())
+                   ->size_bytes_ != CacheSizeUnlimited;
+  }
+  return cache_size_bytes_ != CacheSizeUnlimited;
+}
+
+const LocalCacheSettings* Settings::local_cache_settings() const {
+  return cache_settings_.get();
+}
+
+void Settings::set_local_cache_settings(const LocalCacheSettings& settings) {
+  HARD_ASSERT(persistence_enabled_ == Settings::DefaultPersistenceEnabled,
+              "Cannot set local cache settings, because persistence_enabled "
+              "is already specified. Please remove code specifying "
+              "persistence_enabled.");
+  HARD_ASSERT(
+      cache_size_bytes_ == Settings::DefaultCacheSizeBytes,
+      "Cannot set local cache settings, because cache_size_bytes "
+      "is already specified. Please remove code specifying cache_size_bytes.");
+  cache_settings_ = CopyCacheSettings(settings);
+}
+
+PersistentCacheSettings PersistentCacheSettings::WithSizeBytes(
+    int64_t size) const {
+  PersistentCacheSettings new_settings{*this};
+  new_settings.size_bytes_ = size;
+  return new_settings;
 }
 
 }  // namespace api
