@@ -69,7 +69,7 @@ class CompositeFilter : public Filter {
    * false otherwise.
    */
   bool IsFlatConjunction() const {
-    return IsFlat() && IsConjunction();
+    return composite_filter_rep().IsFlatConjunction();
   }
 
   /**
@@ -115,6 +115,10 @@ class CompositeFilter : public Filter {
     bool IsDisjunction() const;
 
     bool IsFlat() const;
+
+    bool IsFlatConjunction() const {
+      return IsFlat() && IsConjunction();
+    }
 
     bool IsACompositeFilter() const override {
       return true;
