@@ -42,6 +42,7 @@
 #include "Firestore/core/src/model/resource_path.h"
 #include "Firestore/core/src/model/server_timestamp_util.h"
 #include "Firestore/core/src/model/set_mutation.h"
+#include "Firestore/core/src/model/snapshot_version.h"
 #include "Firestore/core/src/model/value_util.h"
 #include "Firestore/core/src/model/verify_mutation.h"
 #include "Firestore/core/src/nanopb/byte_string.h"
@@ -49,6 +50,7 @@
 #include "Firestore/core/src/nanopb/reader.h"
 #include "Firestore/core/src/nanopb/writer.h"
 #include "Firestore/core/src/timestamp_internal.h"
+#include "Firestore/core/src/util/comparison.h"
 #include "Firestore/core/src/util/hard_assert.h"
 #include "Firestore/core/src/util/status.h"
 #include "Firestore/core/src/util/statusor.h"
@@ -111,6 +113,7 @@ using nanopb::SetRepeatedField;
 using nanopb::SharedMessage;
 using nanopb::Writer;
 using remote::WatchChange;
+using util::ComparisonResult;
 using util::ReadContext;
 using util::Status;
 using util::StatusOr;
@@ -635,6 +638,11 @@ google_firestore_v1_Target Serializer::EncodeTarget(
     result.which_resume_type = google_firestore_v1_Target_resume_token_tag;
     result.resume_type.resume_token =
         nanopb::CopyBytesArray(target_data.resume_token().get());
+  } else if (target_data.snapshot_version().CompareTo(
+                 SnapshotVersion::None()) == ComparisonResult::Descending) {
+    result.which_resume_type = google_firestore_v1_Target_read_time_tag;
+    result.resume_type.read_time =
+        EncodeVersion(target_data.snapshot_version());
   }
 
   return result;
@@ -813,8 +821,6 @@ std::vector<Filter> Serializer::DecodeFilters(
 
   // Instead of a singletonList containing AND(F1, F2, ...), we can return
   // a list containing F1, F2, ...
-  // TODO(orquery): Once proper support for composite filters has been
-  // completed, we can remove this flattening from here.
   if (decoded_filter.IsACompositeFilter()) {
     CompositeFilter composite_filter(decoded_filter);
     if (composite_filter.IsFlatConjunction()) {
